Contains programatic segue to home screen from onboarding screen as well as location manager delegate functions
