+++ Section 8/UI/Onboard.storyboard
@@ -10,10 +10,10 @@
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <scenes>
-        <!--View Controller-->
+        <!--onboardingVC-->
         <scene sceneID="Y3e-VK-qdo">
             <objects>
-                <viewController id="sOf-5P-HbW" sceneMemberID="viewController">
+                <viewController storyboardIdentifier="Onboard" title="onboardingVC" id="sOf-5P-HbW" customClass="OnboardingScreenVC" customModule="Section_8" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="T09-zF-uBD">
                         <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
